skills/fabius-cohors/references/agents/python/global-kyc-agent/pyproject.toml
[project]
name = "global-kyc-agent"
version = "0.1.0"
description = "A Global KYC Agent built using Google ADK."
authors = [
    { name = "Syed Gardezi", email = "sgardezi@google.com" },
    { name = "Sandip Datta", email = "sandipdatta@google.com" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk[eval]",
    "companies-house-api-lib==0.4",
    "requests",
    "google-cloud-secret-manager",
    "google-auth",
    "sec-api",
    "beautifulsoup4"
]

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.7",
    "google-adk[eval]>=1.0.0",
    "nest-asyncio>=1.6.0",
    "agent-starter-pack>=0.14.1",
]

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[tool.agent-starter-pack]
example_question = "Perform a compliance check on Google UK Limited."

[tool.agent-starter-pack.settings]
agent_directory = "global_kyc_agent"

[tool.ruff]
# extends the ruff rules established in the root .toml
extend = "../../../pyproject.toml"

[tool.ruff.lint]
# Ignores function complication rule throughout sample
ignore = ["C901", "PLR0915"]

[tool.ruff.lint.per-file-ignores]
# ignores unused import rule exclusively for this sample’s __init__.py files
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["global_kyc_agent"]

